backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.0.0
celery>=5.3.0
redis>=5.0.0
pandas>=2.0.0
openpyxl>=3.0.0
moviepy==1.0.3
Pillow<10.0.0
numpy>=1.24.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.25.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-dotenv>=1.0.0
boto3>=1.28.0  # For S3 storage
minio>=7.1.0   # Alternative to S3